refactor: Extract useApiQuery into separate module
- Create dedicated api/useApiQuery.ts for dynamic API query hook
- Remove useApiQuery implementation from api/hooks.ts
- Import and re-export useApiQuery from the new module
- Maintain full backward compatibility with existing imports
- Improve code organization with single responsibility principle
- Clean separation between core hook logic and convenience wrappers
- Build successful with no compilation errors
